🖥️ui: create and delete justification type actions with cache management
